backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leasecopilot"
version = "0.1.0"
description = "Production-grade RAG for lease contracts (hybrid retrieval + rerank + citations + eval)"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",

  "python-multipart>=0.0.9",

  "langchain>=0.2.0",
  "langchain-community>=0.2.0",
  "langchain-openai>=0.1.0",
  "langchain-cohere>=0.1.0",
  "weaviate-client>=4.6.0",
  "langchain-ollama>=0.1.0",

  "python-dotenv>=1.0.1",

  "pypdf>=4.2.0",
  "python-docx>=1.1.2",

  "ragas>=0.1.15",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
dev-dependencies = [
  "pytest>=8.0.0",
  "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100